refactor: a new utilites folder is created to contain functions used in several components
